+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2018-10-01
+### Added
+- `Referrer-Policy` header rating
+
+## [1.0.2] - 2018-09-14
+### Fixed
+- Bugs in ContentTypeRating when only the `meta` tags are set.
+- Rating of sources and sinks with comments (#41).
+
+### Changed
+- Upgraded `voku/simple_html_dom` to actual version.
+
 ## [1.0.1] - 2018-09-12
 ### Fixed
 - Bugs in ContentTypeRating when only the `meta` tags are set.
@@ -20,5 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CHANGELOG.md and semantic versioning
 
 [Unreleased]: https://github.com/SIWECOS/HSHS-DOMXSS-Scanner/compare/master...development
-<!-- [1.1.0]: https://github.com/SIWECOS/HSHS-DOMXSS-Scanner/compare/1.0.0...1.1.0 -->
+[1.1.0]: https://github.com/SIWECOS/HSHS-DOMXSS-Scanner/compare/1.0.1...1.1.0
+[1.0.2]: https://github.com/SIWECOS/HSHS-DOMXSS-Scanner/compare/1.0.1...1.0.2
+[1.0.1]: https://github.com/SIWECOS/HSHS-DOMXSS-Scanner/compare/1.0.0...1.0.1